thuc hanh 31/1/2024

+++ lab2/src/b6c2slide95.java
@@ -4,7 +4,7 @@ public class b6c2slide95 {
     public static void main(String[] args) {
         Scanner sc = new Scanner (System.in);
         int n;
-        long gthua = 1;
+        double gthua = 1;
         System.out.println("Nhap vao mot so nguyen duong bat ky: ");
         n = sc.nextInt();
         for (int i=1;i<=n;i++){
